--- tests/test_productapi.py
@@ -1,1 +1,0 @@
-"""Test product API/mTLS client things"""
