+++ tests/tlstests/test_productapi.py
@@ -1,0 +1,5 @@
+"""Test product API/mTLS client things"""
+import logging
+
+
+LOGGER = logging.getLogger(__name__)
